LayoutTests/imported/w3c/web-platform-tests/web-animations/responsive/minHeight-expected.txt


PASS minHeight responsive to style changes
FAIL minHeight clamped to 0px assert_equals: expected "0px" but got "7.5px"
